+++ tests/aws/scenario/bookstore/test_bookstore.py
@@ -14,12 +14,12 @@ from aws_cdk.aws_lambda_event_sources import DynamoEventSource
 from constructs import Construct
 
 from localstack.testing.pytest import markers
+from localstack.testing.scenario.cdk_lambda_helper import load_python_lambda_to_s3
 from localstack.testing.scenario.provisioning import InfraProvisioner, cleanup_s3_bucket
 from localstack.testing.snapshots.transformer import GenericTransformer, KeyValueBasedTransformer
 from localstack.utils.files import load_file
 from localstack.utils.strings import to_bytes, to_str
 from localstack.utils.sync import retry
-from tests.aws.cdk_lambda_helper import load_python_lambda_to_s3
 
 if TYPE_CHECKING:
     pass
